子账号测试

+++ app-server/src/main/java/com/wangge/app/server/repository/ChildAccountRepostory.java
@@ -13,7 +13,7 @@ import com.wangge.app.server.entity.ChildAccount;
 * @date 2016年3月25日 下午5:29:02
 *
  */
-public interface ChildAccountRepostory extends JpaRepository<ChildAccount, Integer> {
+public interface ChildAccountRepostory extends JpaRepository<ChildAccount, Long> {
 
   ChildAccount findBySimId(String simId);
